split the huge code simple slam into small pieces
